+++ src/helpers/index.ts
@@ -1,4 +1,4 @@
 export * from './stories';
 export * from './i18n';
 export * from './slotContent';
-export * from './computedClasses';
+export * from './styles';