+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Style variations with `Class Variance Authority`.
+- Actions slot on `SCard` and `SModalCard`.
 
 ## 2.0.0-beta.17 - 2023-10-12
 ### Added